Move signup tests to standalone file.
